% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{edb_sqlite_path}
\alias{edb_sqlite_path}
\title{Default path for the EleccionesDB SQLite snapshot}
\usage{
edb_sqlite_path()
}
\value{
A platform-appropriate file path.
}
\description{
Default path for the EleccionesDB SQLite snapshot
}
